pyproject.toml
[tool.poetry]
name = "regybox"
version = "1.1.0"
description = "Automatically enroll in Regybox Crossfit classes"
authors = ["Martim Lobao <m@rtimlobao.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.11"
beautifulsoup4 = "^4.12.3"
icalendar = "^5.0.11"
python-dotenv = "^1.0.1"
recurring-ical-events = "^2.1.2"
requests = "^2.31.0"
urllib3 = "^2.2.0"

[tool.poetry.group.dev.dependencies]
black = ">=24.1.1"
coverage = ">=7.4.1"
hypothesis = ">=6.97.3"
ipython = ">=8.20.0"
mypy = ">=1.8.0"
pylint = ">=3.0.3"
pytest = ">=8.0.0"
pytest-cov = ">=4.1.0"
ruff = ">=0"
types-beautifulsoup4 = ">=4.12.0.5"
types-pytz = ">=2023.4.0.0"
types-requests = ">=2.31.0.1"
yamllint = ">=1.33.0"

[tool.black]
line_length = 99
preview = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
